requirements.txt
prettytable~=3.3.0
pytest~=7.1.2
requests~=2.28.1
jsonschema~=4.7.2
Faker~=13.15.1
selenium~=4.3.0
allure-python-commons~=2.9.45
assertpy~=1.1